+++ src/octree_partitioner.cpp
@@ -2,1649 +2,690 @@
 #include "openmc/aabb.h"
 #include "openmc/random_dist.h"
 #include "openmc/timer.h"
-#include "openmc/binning.h"
+#include "openmc/partitioner_utils.h"
+#include "openmc/error.h"
 
-#include <stack>
-#include <queue>
-#include <thread>
-#include <algorithm>
-#include <assert.h>
 #include <omp.h>
 
+#include <algorithm>
+#include <stack>
+#include <queue>
 #include <fstream>
 
-#define OCTREE_DEBUG_INFO
+#include <stdlib.h>
 
 namespace openmc {
+    const int32_t OMCP_CURRENT_VERSION[3] = {1, 0, 0};
 
-const int32_t NUM_THREADS = 16;
+    const float   REFINEMENT_SEARCH_DENSITY          = 0.125;
+    const float   REFINEMENT_TIMEOUT                 = 10.0;
+    const int32_t REFINEMENT_GRID_RES                = 128; // ideally should be a power of 2
 
-const int32_t OCTREE_NODE_ALLOCATOR_POOL_SIZE = 16384; // number of 8 children packs
+    const int32_t INFORMATION_REFILLING_START_DEPTH_OFFSET = 1;
 
-const int32_t OMCP_CURRENT_VERSION[3] = {1, 0, 0};
+    enum OMCPStructureType {
+        Octree = 1,
+        KdTree = 2,               // not yet implemented in OMCP
+        ZPlanePartitioner = 3,    // not yet implemented in OMCP
+    };
 
-const float   REFINEMENT_SEARCH_DENSITY = 0.125;
-const float   REFINEMENT_TIMEOUT        = 10.0;
-const int32_t REFINEMENT_GRID_RES       = 128; // ideally should be a power of 2
+    struct OctreeConstructionTask {
+        OctreeUncompressedNode* node;
+        std::vector<CellPoint> points;
 
-const int32_t INFORMATION_REFILLING_START_DEPTH_OFFSET = 1;
+        OctreeConstructionTask() = default;
+        OctreeConstructionTask(OctreeUncompressedNode* n, const std::vector<CellPoint>& p) : node(n), points(p) {}
+    };
 
-const int32_t SKIP_TRAVERSAL_START_DEPTH = 7;
-const int32_t SKIP_TRAVERSAL_GRID_RES = (1 << SKIP_TRAVERSAL_START_DEPTH);
+    OctreeNode::OctreeNode() : data(0) {}
 
-enum OMCPStructureType {
-    Octree = 1,
-    KdTree = 2,               // not yet implemented in OMCP
-    ZPlanePartitioner = 3,    // not yet implemented in OMCP
-};
+    const uint32_t OCTREE_LEAF_FLAG = (1 << 31);
+    void OctreeNode::mark_as_leaf() {
+        data |= OCTREE_LEAF_FLAG;
+    }
 
-OctreeNode::OctreeNode() : data(0) {}
+    bool OctreeNode::is_leaf() const {
+        return (data & OCTREE_LEAF_FLAG);
+    }
 
-const uint32_t OCTREE_LEAF_FLAG = (1 << 31);
-void OctreeNode::mark_leaf() {
-    data |= OCTREE_LEAF_FLAG;
-}
+    void OctreeNode::store_data(uint32_t data) {
+        // remove everything except the flag
+        this->data = (this->data & OCTREE_LEAF_FLAG) | data;
+    }
 
-bool OctreeNode::is_leaf() const {
-    return (data & OCTREE_LEAF_FLAG);
-}
+    uint32_t OctreeNode::read_data() const {
+        return (data & ~OCTREE_LEAF_FLAG);
+    }
 
-void OctreeNode::store_data(uint32_t data) {
-    this->data = data;
-}
+    OctreeUncompressedNode::OctreeUncompressedNode() : children(nullptr), parent(nullptr), depth(0) {}
 
-uint32_t OctreeNode::read_data() const {
-    return (data & ~OCTREE_LEAF_FLAG);
-}
+    bool OctreeUncompressedNode::is_leaf() const {
+        return (children == nullptr);
+    }
 
-class OctreeUncompressedNodeAllocator {
-public:
-  OctreeUncompressedNodeAllocator();
-  ~OctreeUncompressedNodeAllocator();
-  OctreeUncompressedNode* allocate();
-private:
-  std::vector<OctreeUncompressedNode*> pools;
-  int last_pool_next_index;
-};
+    void OctreeUncompressedNode::subdivide() {
+        AABB resultant_boxes[8];
+        resultant_boxes[0] = box;
+        for(int i = 0; i < 3; i++) {
+            AABB temp_box_buffer[8];
 
-struct OctreeConstructionTask {
-    OctreeUncompressedNode* node;
-    std::vector<CellPoint> points;
+            int next_index = 0;
+            for(int idx = 0; idx < (1 << i); idx++) {
+                // split on i-th axis
+                float midpoint = box.get_center()[i];
 
-    OctreeConstructionTask() = default;
-    OctreeConstructionTask(OctreeUncompressedNode* n, const std::vector<CellPoint>& p) : node(n), points(p) {}
-};
+                int j = ((i + 1) % 3);
+                int k = ((i + 2) % 3);
 
-int octree_use_count = 0;
-int fallback_use_count = 0;
-int oob_use_count = 0;
+                AABB splitted_boxes[2] {resultant_boxes[idx], resultant_boxes[idx]};
 
-#ifdef OCTREE_DEBUG_INFO
-omp_lock_t ocm; // short for octree console mutex
-bool ocm_init = false; // unlikely that it will be double init
+                splitted_boxes[0].max[i] = midpoint;
+                splitted_boxes[1].min[i] = midpoint;
 
-// shorten some stuff
-#define LOCK_OCM() omp_set_lock(&ocm)
-#define UNLOCK_OCM() omp_unset_lock(&ocm)
-#endif
+                temp_box_buffer[next_index++] = splitted_boxes[1];
+                temp_box_buffer[next_index++] = splitted_boxes[0];
+            }
 
-OctreeUncompressedNode::OctreeUncompressedNode() : children(nullptr), parent(nullptr), depth(0) {}
-
-bool OctreeUncompressedNode::is_leaf() const {
-    return (children == nullptr);
-}
-
-int OctreeUncompressedNode::get_containing_child_index(const vec3& r) const {
-    auto center = box.get_center();
-    int idx = 4 * int(r.x < center.x) + 2 * int(r.y < center.y) + int(r.z < center.z);
-    return idx;
-}
-
-OctreeUncompressedNode& OctreeUncompressedNode::get_containing_child(const vec3& r) const {
-    return children[get_containing_child_index(r)];
-}
-
-std::vector<AABB> OctreeUncompressedNode::subdivide() {
-    std::vector<AABB> resultant_boxes;
-    resultant_boxes.push_back(box);
-    for(int i = 0; i < 3; i++) {
-        std::vector<AABB> temp_box_buffer;
-
-        for(const auto& box : resultant_boxes) {
-            // split on i-th axis
-            float midpoint = box.get_center()[i];
-
-            int j = ((i + 1) % 3);
-            int k = ((i + 2) % 3);
-
-            AABB splitted_boxes[2];
-            vec3 extension_point[2];
-
-            extension_point[0][i] = midpoint;
-            extension_point[0][j] = box.min[j];
-            extension_point[0][k] = box.min[k];
-
-            extension_point[1][i] = midpoint;
-            extension_point[1][j] = box.max[j];
-            extension_point[1][k] = box.max[k];
-
-            splitted_boxes[0].extend(box.max);
-            splitted_boxes[0].extend(extension_point[0]);
-
-            splitted_boxes[1].extend(box.min);
-            splitted_boxes[1].extend(extension_point[1]);
-
-            temp_box_buffer.push_back(splitted_boxes[0]);
-            temp_box_buffer.push_back(splitted_boxes[1]);
+            // move the results to the next splitting stage
+            std::copy(temp_box_buffer, temp_box_buffer + (2 << i), resultant_boxes);
         }
-
-        // move the results to the next splitting stage
-        resultant_boxes = temp_box_buffer;
-    }
-
-    for(int i = 0; i < 8; i++) {
-        children[i].box = resultant_boxes[i];
-    }
-
-    return resultant_boxes;
-}
- 
-bool OctreeUncompressedNode::contains(int cell) const {
-    return std::binary_search(cells.begin(), cells.end(), cell);
-}
-
-OctreeUncompressedNodeAllocator::OctreeUncompressedNodeAllocator() : last_pool_next_index(0) {}
-
-OctreeUncompressedNodeAllocator::~OctreeUncompressedNodeAllocator() {
-    for(auto* ptr : pools) {
-        delete[] ptr;
-    }
-}
-
-OctreeUncompressedNode* OctreeUncompressedNodeAllocator::allocate() {
-    if(last_pool_next_index == OCTREE_NODE_ALLOCATOR_POOL_SIZE || pools.size() == 0) {
-        pools.push_back(new OctreeUncompressedNode[8 * OCTREE_NODE_ALLOCATOR_POOL_SIZE]);
-        last_pool_next_index = 0;
-    }
-
-    auto ptr = &pools.back()[8 * last_pool_next_index];
-    last_pool_next_index++;
-    return ptr;
-}
-
-void pick_untested_cells(const std::vector<int>& tested_cells, const std::vector<int>& possible_cells, std::vector<int>& untested_cells) {
-    untested_cells.clear();
-    int next_idx = 0;
-    for(int cell : possible_cells) {
-        if(next_idx < tested_cells.size() && tested_cells[next_idx] == cell) {
-            next_idx++;
-        } else {
-            untested_cells.push_back(cell);
-        }
-    }
-}
-
-std::vector<OctreeUncompressedNode*> store_children(OctreeUncompressedNode& root) {
-    std::vector<OctreeUncompressedNode*> children;
-    children.reserve(1000000);
-
-    std::stack<OctreeUncompressedNode*> unproc_nodes;
-    unproc_nodes.push(&root);
-    while(!unproc_nodes.empty()) {
-        auto cur = unproc_nodes.top();
-        unproc_nodes.pop();
 
         for(int i = 0; i < 8; i++) {
-            auto ptr = &cur->children[i];
-            if(ptr->is_leaf()){
-                children.push_back(ptr);
+            children[i].box = resultant_boxes[i];
+            children[i].depth = depth + 1;
+            children[i].parent = this;
+        }
+    }
+    
+    bool OctreeUncompressedNode::contains(int cell) const {
+        return std::binary_search(cells.begin(), cells.end(), cell);
+    }
+
+    void pick_untested_cells(const std::vector<int>& tested_cells, const std::vector<int>& possible_cells, std::vector<int>& untested_cells) {
+        untested_cells.clear();
+        int next_idx = 0;
+        for(int cell : possible_cells) {
+            if(next_idx < tested_cells.size() && tested_cells[next_idx] == cell) {
+                next_idx++;
             } else {
-                unproc_nodes.push(ptr);
+                untested_cells.push_back(cell);
             }
         }
     }
 
-    std::cout << "There are " << children.size() << " children nodes in the octree.\n";
-    std::cout.flush();
+    void refine_octree_random(
+        const Universe& univ,
+        const UniversePartitioner& fallback,
+        const AABB& bounds,
+        const std::vector<OctreeUncompressedNode*>& leaves
+    ) {
+        const int32_t num_threads = omp_get_max_threads();
 
-    return children;
-}
-
-void refine_octree_random(
-    const Universe& univ,
-    const UniversePartitioner& fallback,
-    OctreeUncompressedNodeAllocator& node_alloc,
-    OctreeUncompressedNode& root
-) {
-    Timer refinement_timer;
-    refinement_timer.start();
-    float search_time = 0.0, insert_time = 0.0;
-
-    std::cout << "Refining octree via random refinement..\n";
-    std::cout.flush();
-
-    std::vector<uint64_t[2]> rng_node_selec(NUM_THREADS);
-    std::vector<uint64_t[3]> rng_pos(NUM_THREADS);
-
-    for(int i = 0; i < NUM_THREADS; i++) {
-        rng_node_selec[i][0] = i;
-        rng_node_selec[i][1] = i + 324 * NUM_THREADS;
-        for(int j = 0; j < 3; j++) {
-            rng_pos[i][j] = NUM_THREADS * (j + 1) + i;
-        }
-    }
-
-    auto children = store_children(root);
-    std::vector<omp_lock_t> cell_locks(children.size());
-
-    for(auto& lock : cell_locks) {
-        omp_init_lock(&lock);
-    }
-
-    omp_lock_t allocate_lock, replace_lock;
-    omp_init_lock(&allocate_lock);
-    omp_init_lock(&replace_lock);
-
-    struct ProbabilityBin {
-    public:
-        ProbabilityBin() : num_searched(0), num_found(0) {
-            //common_cells.reserve(200);
-            //found_cells.reserve(500);
-
-            omp_init_lock(&found_lock);
+        // generate the seeds
+        std::vector<uint64_t[2]> rng_node_selec(num_threads);
+        std::vector<uint64_t[3]> rng_pos(num_threads);
+        for(int i = 0; i < num_threads; i++) {
+            rng_node_selec[i][0] = i;
+            rng_node_selec[i][1] = i + num_threads;
+            for(int j = 0; j < 3; j++) {
+                rng_pos[i][j] = i + num_threads * (j + 2);
+            }
         }
 
-        ~ProbabilityBin() {
-            omp_destroy_lock(&found_lock);
+        using ProbBinT = ProbabilityBin<OctreeUncompressedNode>;
+        std::vector<ProbBinT> prob_bin_grid(REFINEMENT_GRID_RES * REFINEMENT_GRID_RES * REFINEMENT_GRID_RES);
 
-            for(auto& p : contained_nodes) {
-                omp_destroy_lock(&p.second);
+        vec3 prob_bin_dim;
+        for(int i = 0; i < 3; i++) {
+            prob_bin_dim[i] = (bounds.max[i] - bounds.min[i]) / REFINEMENT_GRID_RES;
+        }
+
+        for(auto leaf : leaves) {
+            int idx = 0;
+            for(int i = 0; i < 3; i++) {
+                idx = REFINEMENT_GRID_RES * idx + int( (leaf->box.get_center()[i] - bounds.min[i]) / prob_bin_dim[i] );
             }
 
-            common_cells.clear();
-            found_cells.clear();
-
-            common_cells.shrink_to_fit();
-            found_cells.shrink_to_fit();
+            omp_lock_t dummy;
+            prob_bin_grid[idx].contained_nodes.emplace_back(leaf, dummy);
+            for(int cell : leaf->cells) {
+                prob_bin_grid[idx].found_cells.push_back(cell);
+            }
         }
 
-        float compute_score() {
-            const float K_FOUND_MULT = 1.0;
-            const float K_SEARCHED_MULT = 1.0;
-
-            float found_score = std::max(K_FOUND_MULT * num_found, 1.0f) / std::max(K_SEARCHED_MULT * num_searched, 1.0f);
-            found_score *= found_score;
-
-            float size_score = contained_nodes.size();
-            return size_score;
+        for(auto& prob_bin : prob_bin_grid) {
+            prob_bin.update_common_cells();
+            for(auto& p : prob_bin.contained_nodes) {
+                omp_init_lock(&p.second);
+            }
         }
 
-        int num_searched, num_found;
-        std::vector<std::pair<OctreeUncompressedNode*, omp_lock_t>> contained_nodes;
-        //std::vector<float> node_cdf;
+        std::vector<float> bin_cdf(prob_bin_grid.size());
 
-        /*
-        void update_cdf() {
+        Timer timeout_timer;
+        timeout_timer.start();
+        int iteration = 0;
+        while(timeout_timer.elapsed() < REFINEMENT_TIMEOUT) {
+            int num_search_points = (int)(bounds.volume() * REFINEMENT_SEARCH_DENSITY);
+            int num_points_searched = 0;
+
+            // first, generate cdf
             float total_cdf = 0.0;
-            for(int i = 0; i < node_cdf.size(); i++) {
-                float score = contained_nodes[i].first->cells.size();
-                total_cdf += score;
-                node_cdf[i] = total_cdf;
+            for(int i = 0; i < prob_bin_grid.size(); i++) {
+                total_cdf += prob_bin_grid[i].compute_score();
+                bin_cdf[i] = total_cdf;
             }
 
-            for(float& cdf : node_cdf) {
+            for(float& cdf : bin_cdf) {
                 cdf /= total_cdf;
             }
-        }
-        */
-
-        size_t pick_node(uint64_t* seed) {
-            size_t rand_idx = (size_t)uniform_distribution(0.0, (double)contained_nodes.size(), seed);
-            return rand_idx;
-            /*
-            float cdf_val = uniform_distribution(0.0, 1.0, seed);
-            auto cdf_iter = std::upper_bound(node_cdf.begin(), node_cdf.end(), cdf_val);
-            size_t node_idx = std::distance(node_cdf.begin(), cdf_iter);
-            if(node_idx == node_cdf.size()) {
-                node_idx--;
-            }     
-
-            return node_idx;   
-            */
-        }
-
-        void update_common_cells() {
-            found_cells.reserve(common_cells.size() + found_cells.size());
             
-            for(int cell : common_cells) {
-                found_cells.push_back(cell);
-            }
+            #pragma omp parallel for
+            for(int tid = 0; tid < num_threads; tid++) {
+                std::vector<int> untested_cells, temp_buf;
+                untested_cells.reserve(256);
+                temp_buf.reserve(256);
 
-            std::sort(found_cells.begin(), found_cells.end());
+                // we don't need a lock for loop iteration logic since nothing bad will happen if race conditions occur
+                // maybe we do a few points less or more, but that doesn't matter
+                while(num_points_searched < num_search_points) {
+                    num_points_searched++;
 
-            common_cells.clear();
-
-            int prev_cell = -1;
-            for(int cell : found_cells) {
-                if(cell != prev_cell) {
-                    common_cells.push_back(cell);
-                    prev_cell = cell;
-                }
-            }
-
-            found_cells.clear();
-        }
-
-        void add_cell(int cell) {
-            omp_set_lock(&found_lock);
-            found_cells.push_back(cell);
-            omp_unset_lock(&found_lock);
-        }
-
-        omp_lock_t found_lock;
-        std::vector<int> common_cells;
-        std::vector<int> found_cells;
-    };
-
-    std::cout << "Allocating probabiliy bins..." << std::endl;
-
-    std::vector<ProbabilityBin> prob_bin_grid(REFINEMENT_GRID_RES * REFINEMENT_GRID_RES * REFINEMENT_GRID_RES);
-
-
-    vec3 prob_bin_dim;
-    for(int i = 0; i < 3; i++) {
-        prob_bin_dim[i] = (root.box.max[i] - root.box.min[i]) / REFINEMENT_GRID_RES;
-    }
-
-    for(auto child : children) {
-        int idx3d[3];
-        for(int i = 0; i < 3; i++) {
-            idx3d[i] = (child->box.get_center()[i] - root.box.min[i]) / prob_bin_dim[i];
-        }
-
-        int idx = 
-            idx3d[0] +
-            idx3d[1] * REFINEMENT_GRID_RES +
-            idx3d[2] * REFINEMENT_GRID_RES * REFINEMENT_GRID_RES;
-
-        omp_lock_t dummy;
-        prob_bin_grid[idx].contained_nodes.emplace_back(child, dummy);
-        for(int cell : child->cells) {
-            prob_bin_grid[idx].found_cells.push_back(cell);
-        }
-    }
-
-
-    for(auto& prob_bin : prob_bin_grid) {
-        //prob_bin.node_cdf.resize(prob_bin.contained_nodes.size());
-        prob_bin.update_common_cells();
-        for(auto& p : prob_bin.contained_nodes) {
-            omp_init_lock(&p.second);
-        }
-    }
-
-    std::cout << "Done allocating probability bins!" << std::endl;
-
-    std::vector<float> bin_cdf(prob_bin_grid.size());
-
-    Timer timeout_timer;
-    timeout_timer.start();
-
-    int iteration = 0;
-
-    while(timeout_timer.elapsed() < REFINEMENT_TIMEOUT) {
-        int num_search_points = (int)(root.box.volume() * REFINEMENT_SEARCH_DENSITY);
-        int num_points_searched = 0;
-        int num_total_missing_cells = 0;
-
-        // first, generate cdf
-        // std::cout << "GENERATING CDF..." << std::endl;
-        float total_cdf = 0.0;
-        for(int i = 0; i < prob_bin_grid.size(); i++) {
-            // compute score
-            float score = prob_bin_grid[i].compute_score();
-
-            total_cdf += score;
-            bin_cdf[i] = total_cdf;
-        }
-
-        for(float& cdf : bin_cdf) {
-            cdf /= total_cdf;
-        }
-        
-        //for(auto& prob_bin : prob_bin_grid) {
-        //    prob_bin.update_cdf();
-        //}
-
-        //std::cout << "SEARCHING POINTS..." << std::endl;
-        Timer t;
-        t.start();
-        #pragma omp parallel for
-        for(int tid = 0; tid < NUM_THREADS; tid++) {
-            std::vector<int> untested_cells, temp_buf;
-            untested_cells.reserve(256);
-            temp_buf.reserve(256);
-
-            // we don't need a lock for loop iteration logic since nothing bad will happen if race conditions occur
-            // maybe we do a few points less or more, but that doesn't matter
-            while(num_points_searched < num_search_points) {
-                num_points_searched++;
-
-                ProbabilityBin* prob_bin;
-                do {
-                    float cdf_val = uniform_distribution(0.0, 1.0, &rng_node_selec[tid][0]);
-                    auto cdf_iter = std::upper_bound(bin_cdf.begin(), bin_cdf.end(), cdf_val);
-                    size_t bin_idx = std::distance(bin_cdf.begin(), cdf_iter);
-                    if(bin_idx == bin_cdf.size()) {
-                        bin_idx--;
-                    }
-
-                    prob_bin = &prob_bin_grid[bin_idx];
-                } while(prob_bin->contained_nodes.size() == 0);
-                prob_bin->num_searched++;
-
-                size_t idx = prob_bin->pick_node(&rng_node_selec[tid][1]);
-
-                // in the rare case some wierd floating point stuff happens
-                if(idx >= prob_bin->contained_nodes.size()) {
-                    idx =  prob_bin->contained_nodes.size() - 1;
-                }
-
-                OctreeUncompressedNode* current = prob_bin->contained_nodes[idx].first;
-                omp_lock_t* cell_lock = &prob_bin->contained_nodes[idx].second;
-
-                CellPointUncompressed point;
-                for(int i = 0; i < 3; i++) {
-                    point.pos[i] = uniform_distribution(current->box.min[i], current->box.max[i], &rng_pos[tid][i]);
-                }
-
-
-                omp_set_lock(cell_lock);
-                point.cell = univ.find_cell_for_point(current->cells, point.pos);
-                omp_unset_lock(cell_lock);
-
-                if(point.cell == -1) {
-                    prob_bin->num_found++;
-
-                    pick_untested_cells(current->cells, prob_bin->common_cells, untested_cells);
-                    point.cell = univ.find_cell_for_point(untested_cells, point.pos);
-                    if(point.cell == -1) {
-                        if(current->parent) {
-                            point.cell = univ.find_cell_for_point(current->parent->cells, point.pos);
+                    ProbBinT* prob_bin;
+                    do {
+                        float cdf_val = uniform_distribution(0.0, 1.0, &rng_node_selec[tid][0]);
+                        auto cdf_iter = std::upper_bound(bin_cdf.begin(), bin_cdf.end(), cdf_val);
+                        size_t bin_idx = std::distance(bin_cdf.begin(), cdf_iter);
+                        if(bin_idx == bin_cdf.size()) {
+                            bin_idx--;
                         }
 
-                        if(point.cell == -1) {
-                            Direction dummy{0, 0, 1};
-                            const auto& possible_cells = fallback.get_cells(point.pos, dummy);
+                        prob_bin = &prob_bin_grid[bin_idx];
+                    } while(prob_bin->contained_nodes.size() == 0);
+                    prob_bin->num_searched++;
 
-                            pick_untested_cells(untested_cells, possible_cells, temp_buf);
-
-                            omp_set_lock(cell_lock);
-                            pick_untested_cells(current->cells, temp_buf, untested_cells);
-                            omp_unset_lock(cell_lock);
-
-                            point.cell = univ.find_cell_for_point(untested_cells, point.pos);
-
-                            // rarely happens, but when it does, it segfaults the program
-                            if(point.cell == -1) {
-                                point.cell = univ.find_cell_for_point(univ.cells_, point.pos);
-                            }
-                        }
-
-                        prob_bin->add_cell(point.cell);
+                    size_t idx = prob_bin->pick_node(&rng_node_selec[tid][1]);
+                    if(idx >= prob_bin->contained_nodes.size()) {
+                        idx =  prob_bin->contained_nodes.size() - 1;
                     }
 
+                    OctreeUncompressedNode* current = prob_bin->contained_nodes[idx].first;
+                    omp_lock_t* cell_lock = &prob_bin->contained_nodes[idx].second;
+
+                    CellPointUncompressed point;
+                    for(int i = 0; i < 3; i++) {
+                        point.pos[i] = uniform_distribution(current->box.min[i], current->box.max[i], &rng_pos[tid][i]);
+                    }
 
                     omp_set_lock(cell_lock);
-                    current->cells.push_back(point.cell);
-                    std::sort(current->cells.begin(), current->cells.end());
+                    point.cell = univ.find_cell_for_point(current->cells, point.pos);
                     omp_unset_lock(cell_lock);
 
-                    num_total_missing_cells++;
+                    if(point.cell == -1) {
+                        prob_bin->num_found++;
+
+                        omp_set_lock(cell_lock);
+                        pick_untested_cells(current->cells, prob_bin->common_cells, untested_cells);
+                        omp_unset_lock(cell_lock);
+
+                        point.cell = univ.find_cell_for_point(untested_cells, point.pos);
+                        if(point.cell == -1) {
+                            point.cell = univ.find_cell_for_point(current->parent->cells, point.pos);
+
+                            if(point.cell == -1) {
+                                Direction dummy{0, 0, 1};
+                                const auto& possible_cells = fallback.get_cells(point.pos, dummy);
+
+                                pick_untested_cells(untested_cells, possible_cells, temp_buf);
+
+                                omp_set_lock(cell_lock);
+                                pick_untested_cells(current->cells, temp_buf, untested_cells);
+                                omp_unset_lock(cell_lock);
+
+                                point.cell = univ.find_cell_for_point(untested_cells, point.pos);
+
+                                // very rarely, even the fallback misses a cell
+                                // we need to do an exhaustive search or the program will segfault
+                                if(point.cell == -1) {
+                                    point.cell = univ.find_cell_for_point(univ.cells_, point.pos);
+                                }
+                            }
+
+                            prob_bin->add_cell(point.cell);
+                        }
+
+
+                        omp_set_lock(cell_lock);
+                        // insertion sort
+                        current->cells.push_back(point.cell);
+                        for(int i = current->cells.size() - 1; i > 0; i--) {
+                            if(current->cells[i] < current->cells[i - 1]) {
+                                std::swap(current->cells[i], current->cells[i - 1]);
+                            }
+                        }
+                        omp_unset_lock(cell_lock);
+                    }
                 }
             }
-        }
-
-        search_time += t.elapsed();
-        t.reset();
-
-        std::cout << "Refinement on iteration " << (iteration + 1) << ":\t" << num_search_points << "\tcells searched,\t" << num_total_missing_cells << "\tmising cells inserted into octree. (" << (100.0 * num_total_missing_cells) / num_search_points << "%\tdiscovery rate)\n";
-        std::cout.flush();
-
-        iteration++;
-    }
-
-    float total = refinement_timer.elapsed();
-    std::cout << "Refinement took " << total << " seconds, here's the breakdown:\n";
-    std::cout << "\tSearch:\t" << search_time << "\tseconds\n";
-    std::cout << "\tInsert:\t" << insert_time << "\tseconds\n";
-    std::cout << "\tExpect:\t" << search_time + insert_time << "\tseconds\n";
-    std::cout.flush();
-
-    for(auto& lock : cell_locks) {
-        omp_destroy_lock(&lock);
-    }
-
-    prob_bin_grid.clear();
-    prob_bin_grid.shrink_to_fit();
-}
-
-float tot_cons_time;
-OctreePartitioner::OctreePartitioner(const Universe& univ, int target_cells_per_node, int max_depth, const std::string& file_path) : fallback(univ) {
-    if(!ocm_init) {
-        ocm_init = true;
-        omp_init_lock(&ocm);
-    }
-
-    Timer t;
-    t.start();
-    std::cout << "====================OCTREE CONSTRUCTION===================\n";
-
-    max_depth = 999;
-    float depth_vh_mult[] = {
-        1.0, 1.0, 1.0, 1.5, 2.5, 4.0,
-        6.0, 12.5, 19.0, 32.0, 64.0, 128.0, 999.0
-    };
-
-    int max_tree_size = (2 << (3 * max_depth));
-    std::cout << "Maximum tree size is " << max_tree_size << '\n';
-    // octrees don't have to have cube nodes/leaves, so we don't have to convert the source AABB into a square
-    auto urootbox = univ.bounding_box();
-    AABB rootbox(vec3(urootbox.xmin, urootbox.ymin, urootbox.zmin), vec3(urootbox.xmax, urootbox.ymax, urootbox.zmax));
-
-    const float half_side_length = 130.0;
-    rootbox.min = vec3(-half_side_length, -half_side_length, -half_side_length);
-    rootbox.max = vec3( half_side_length,  half_side_length,  half_side_length);
-
-    bounds = rootbox;
-    OctreeUncompressedNode root;
-    root.depth = 0;
-    root.id = 0;
-    root.box = bounds;
-
-    if(univ.cells_.size() <= target_cells_per_node) {
-        std::cout << "Universe has only " << root.cells.size() << " cells, which is below the target cells per node, " 
-                  << target_cells_per_node << ". Octree will have depth of only 1.\n";
-        root.cells = univ.cells_;
-        return;
-    }
-
-    Timer binning_timer;
-    binning_timer.start();
-
-    auto points_in_bounds = binned_point_search<CellPoint>(univ, fallback, bounds);
-    binning_timer.stop();
-    std::cout << "Done searching for points! Point search took " << binning_timer.elapsed() << " seconds. Beginning organization of points in octree...\n";
-    std::cout.flush();
-
-    // kind of a crude way to sort but whatever
-    Timer sorting_timer;
-    sorting_timer.start();
-   
-    struct PointComp {
-        inline bool operator() (const CellPoint& lhs, const CellPoint& rhs) {
-            return (lhs.get_cell() < rhs.get_cell());
-        }
-    }; 
-    std::sort(points_in_bounds.begin(), points_in_bounds.end(), PointComp());
-
-    int prev_cell = -1;
-    for(const auto& p : points_in_bounds) {
-        if(prev_cell != p.get_cell()) {
-            root.cells.push_back(p.get_cell());
-            prev_cell = p.get_cell();
+            iteration++;
         }
     }
 
-    root.num_unique_cells = root.cells.size();
+    OctreePartitioner::OctreePartitioner(const Universe& univ, int target_cells_per_node) : fallback(univ) {
+        const float half_side_length = 130.0;
+        bounds.min = vec3(-half_side_length, -half_side_length, -half_side_length);
+        bounds.max = vec3( half_side_length,  half_side_length,  half_side_length);
 
-    sorting_timer.stop();
-    std::cout << "Total sorting time: " << sorting_timer.elapsed() << '\n';
-    std::cout.flush();
-
-    double popping_time = 0.0, init_time = 0.0, alloc_time = 0.0, node_selec_time = 0.0, post_process_time = 0.0;
-
-    num_nodes = 1;
-    int next_id = 1; // root is already 0
-    int num_leaves = 0;
-    uint32_t max_encountered_depth = 0;
-    uint32_t min_leaf_depth = UINT32_MAX;
-
-    std::queue<OctreeConstructionTask> oct_unproc;
-    oct_unproc.emplace(&root, points_in_bounds);
-    OctreeUncompressedNodeAllocator node_alloc;
-    std::vector<OctreeUncompressedNode*> nodes_to_propagate;
-    nodes_to_propagate.push_back(&root);
-    Timer total_building_timer;
-    total_building_timer.start();
-    while(!oct_unproc.empty()) {
-        Timer t;
-
-        t.start(); 
-        auto cur_task = std::move(oct_unproc.front());
-        oct_unproc.pop();
-        popping_time += t.elapsed();
-        t.reset();
-
-        max_encountered_depth = std::max(max_encountered_depth, (uint32_t)(cur_task.node->depth + 1)); 
-
-        // subdivide
-        t.start();
-
-        cur_task.node->children = node_alloc.allocate();
-
-        auto boxes = cur_task.node->subdivide();
-        init_time += t.elapsed();
-        t.reset();
-
-        // allocation
-        t.start();
-        OctreeConstructionTask child_tasks[8];
-        for(int i = 0; i < 8; i++) {
-            child_tasks[i].node = &cur_task.node->children[i];
-            child_tasks[i].node->box = boxes[i];
-            //child_tasks[i].points.reserve(cur_task.points.size());
-
-            child_tasks[i].node->depth = cur_task.node->depth + 1;
-            child_tasks[i].node->id = next_id++;
-
-            child_tasks[i].node->parent = cur_task.node;
-            nodes_to_propagate.push_back(child_tasks[i].node);
+        if(univ.cells_.size() <= target_cells_per_node) {
+            warning("Universe has only "  + std::to_string(univ.cells_.size()) + 
+                    " cells, which is below the target cells per node, which is " 
+                    + std::to_string(target_cells_per_node) + " cells. Octree will only consist of root.");
+            
+            OctreeNode node;
+            node.store_data(0);
+            node.mark_as_leaf();
+            cell_data.push_back(univ.cells_);            
+            return;
         }
-        alloc_time += t.elapsed();
-        t.reset();
 
-        // sort points
-        t.start();
-        for(const auto& point : cur_task.points) {
-            child_tasks[point.get_child_index(cur_task.node->depth)].points.push_back(point);
-        }
-        node_selec_time += t.elapsed();
-        t.reset();
+        write_message("Building octree...", 5);
 
-        float parent_vh = cur_task.node->box.volume() * cur_task.node->num_unique_cells;
-        float children_vh = 0.0;
+        Timer construction_timer;
+        construction_timer.start();
 
-        // post processing (make nodes leaves or push on construction stack)
-        t.start();
-        bool force_subdiv = false;
-        for(int i = 0; i < 8; i++) {
-            // we don't have to sort since all points were already sorted before processing loop
-            int prev_cell = -1;
-            int num_unique_cells = 0;
-            for(const auto& p : child_tasks[i].points) {
-                if(p.get_cell() != prev_cell) {
-                    num_unique_cells++;
-                    prev_cell = p.get_cell();
-                }
-            }
+        auto points_in_bounds = binned_point_search<CellPoint>(univ, fallback, bounds);
 
-            const int NUM_UNIQUE_CELLS_MULT = 1; // prevent refinement stage from having to reallocate
-            child_tasks[i].node->cells.reserve(NUM_UNIQUE_CELLS_MULT * num_unique_cells);
-            child_tasks[i].node->num_unique_cells = num_unique_cells;
-
-            children_vh += child_tasks[i].node->box.volume() * num_unique_cells;
-
-            if(num_unique_cells > target_cells_per_node) {
-                force_subdiv = true;
+        OctreeUncompressedNode root;
+        root.box = bounds;
+        root.depth = 0;
+    
+        std::sort(points_in_bounds.begin(), points_in_bounds.end());
+        int prev_cell = -1;
+        for(const auto& p : points_in_bounds) {
+            if(prev_cell != p.get_cell()) {
+                root.cells.push_back(p.get_cell());
+                prev_cell = p.get_cell();
             }
         }
+        root.num_unique_cells = root.cells.size();
 
-        if(force_subdiv || depth_vh_mult[cur_task.node->depth] * children_vh < parent_vh) {
-            num_nodes += 8;
+        num_nodes = 1;
+        num_leaves = 0;
+
+        float depth_vh_mult[] = {
+            1.0, 1.0, 1.0, 1.5, 2.5, 4.0,
+            6.0, 12.5, 19.0, 32.0, 64.0, 128.0, 
+            999.0, 9999.0, 99999.0
+        };
+
+        NodeAllocator<OctreeUncompressedNode, 8> node_alloc;
+        std::vector<OctreeUncompressedNode*> nodes_to_propagate {&root};
+        std::vector<OctreeUncompressedNode*> leaves;
+
+        // this section of code still needs to be multithreaded
+        // it can become a bottleneck, espcially with large number of points
+        std::queue<OctreeConstructionTask> unprocessed_tasks;
+        unprocessed_tasks.emplace(&root, points_in_bounds);
+        while(!unprocessed_tasks.empty()) {
+
+            auto cur_task = std::move(unprocessed_tasks.front());
+            unprocessed_tasks.pop();
+
+            // subdivide
+            cur_task.node->children = node_alloc.allocate();
+            cur_task.node->subdivide();
+
+            OctreeConstructionTask child_tasks[8];
             for(int i = 0; i < 8; i++) {
-                prev_cell = -1;
+                child_tasks[i].node = &cur_task.node->children[i];
+                nodes_to_propagate.push_back(child_tasks[i].node);
+            }
+
+
+            // sort points
+            for(const auto& point : cur_task.points) {
+                child_tasks[point.get_child_index(cur_task.node->depth)].points.push_back(point);
+            }
+
+            float parent_vh = cur_task.node->box.volume() * cur_task.node->num_unique_cells;
+            float children_vh = 0.0;
+
+            // post processing (make nodes leaves or push on construction stack)
+            bool force_subdiv = false;
+            for(int i = 0; i < 8; i++) {
+                // count the number of unique cells
+                int num_unique_cells = 0;
+                int prev_cell = -1;
                 for(const auto& p : child_tasks[i].points) {
                     if(p.get_cell() != prev_cell) {
-                        child_tasks[i].node->cells.push_back(p.get_cell());
+                        num_unique_cells++;
                         prev_cell = p.get_cell();
                     }
                 }
 
-                //child_tasks[i].node->cells.reserve(num_unique_cells * 2);
-                oct_unproc.push(std::move(child_tasks[i]));
-            }
-        } else {
-            cur_task.node->children = nullptr; // mark as leaf
-            num_leaves++;
+                child_tasks[i].node->num_unique_cells = num_unique_cells;
 
-            min_leaf_depth = std::min(min_leaf_depth, (uint32_t)cur_task.node->depth);
-        }
-
-        post_process_time += t.elapsed();
-
-        // free memory
-        cur_task.points.clear();
-        cur_task.points.shrink_to_fit();
-
-        t.reset();
-    }
-    total_building_timer.stop();
-
-    std::cout << "Max encountered depth is " << max_encountered_depth << std::endl;
-    std::cout << "Min leaf depth is " << min_leaf_depth << std::endl;
-
-    std::cout << "Done placing points into octree! Performance statistics:\n";
-    std::cout << "\tTask popping:\t" << popping_time << '\n';
-    std::cout << "\tChild init:\t" << init_time << '\n';
-    std::cout << "\tAllocation:\t" << alloc_time << '\n';
-    std::cout << "\tNode selec:\t" << node_selec_time << '\n';
-    std::cout << "\tPost proc:\t" << post_process_time << '\n';
-    std::cout << "\tExpec tot:\t" << popping_time + init_time + alloc_time + node_selec_time + post_process_time << '\n';
-    std::cout << "\tTotal time:\t" << total_building_timer.elapsed() << '\n';
-    std::cout.flush();
-
-    refine_octree_random(univ, fallback, node_alloc, root);
-
-    // now, build the cells list
-    Timer node_propagation_timer;
-    node_propagation_timer.start();
-    struct NodeComp {
-        inline bool operator()(const OctreeUncompressedNode* lhs, const OctreeUncompressedNode* rhs) const {
-            return (lhs->depth < rhs->depth);
-        }
-    };
-    std::sort(nodes_to_propagate.rbegin(), nodes_to_propagate.rend(), NodeComp());
-
-    for(auto ptr : nodes_to_propagate) {
-        auto& cur = *ptr;
-
-        if(!cur.is_leaf()) {
-            std::sort(cur.cells.begin(), cur.cells.end());
-
-            int prev_cell = cur.cells[0];
-
-            int next_idx = 1;
-            for(int i = 1; i < cur.cells.size(); i++) {
-                if(cur.cells[i] != prev_cell) {
-                    cur.cells[next_idx] = cur.cells[i];
-                    next_idx++;
-                    prev_cell = cur.cells[i];
+                children_vh += child_tasks[i].node->box.volume() * num_unique_cells;
+                if(num_unique_cells > target_cells_per_node) {
+                    force_subdiv = true;
                 }
             }
-            cur.cells.resize(next_idx);
-            cur.cells.shrink_to_fit();
-        }
 
-        if(cur.parent) {
-            for(int cell : cur.cells) {
-                cur.parent->cells.push_back(cell);
-            }
-        }
-    }
-    node_propagation_timer.stop();
-    std::cout << "Node propagation took " << node_propagation_timer.elapsed() << " seconds." << std::endl;
-
-
-    Timer information_refilling_timer;
-    information_refilling_timer.start();
-    // a possible way to remove issues with missing information in nodes
-    std::stack<OctreeUncompressedNode*> uninfo_refilled_nodes;
-    uninfo_refilled_nodes.push(&root);
-    while(!uninfo_refilled_nodes.empty()) {
-        auto cur = uninfo_refilled_nodes.top();
-        uninfo_refilled_nodes.pop();
-
-        bool should_collect_leaves = false;
-        for(int i = 0; i < 8; i++) {
-            if(cur->children[i].is_leaf()) {
-                should_collect_leaves = true;
-                break;
-            }
-        }
-        
-        auto collection_start = cur;
-        #if 1
-        // faster but worse results
-        for(int i = 0; i < INFORMATION_REFILLING_START_DEPTH_OFFSET; i++) {
-            if(collection_start->parent) {
-                collection_start = collection_start->parent;
-            }
-        }
-        
-        const auto& unique_cells = collection_start->cells;
-
-        for(int i = 0; i < 8; i++) {
-            if(cur->children[i].is_leaf()) {
-                int orig_size = cur->children[i].cells.size();
-
-                auto search_start = cur->children[i].cells.begin();
-                auto search_end   = cur->children[i].cells.begin() + orig_size;
-
-                cur->children[i].num_original_cells = orig_size;
-
-                for(int cell : unique_cells) {
-                    if(!std::binary_search(search_start, search_end, cell)) {
-                        cur->children[i].cells.push_back(cell);
+            if(force_subdiv || depth_vh_mult[cur_task.node->depth] * children_vh < parent_vh) {
+                // continue subdivision on this branch
+                num_nodes += 8;
+                for(int i = 0; i < 8; i++) {
+                    child_tasks[i].node->cells.reserve(child_tasks[i].node->num_unique_cells);
+                    prev_cell = -1;
+                    for(const auto& p : child_tasks[i].points) {
+                        if(p.get_cell() != prev_cell) {
+                            child_tasks[i].node->cells.push_back(p.get_cell());
+                            prev_cell = p.get_cell();
+                        }
                     }
+
+                    unprocessed_tasks.push(std::move(child_tasks[i]));
                 }
             } else {
-                uninfo_refilled_nodes.push(&cur->children[i]);
-            }
-        }
-        #else
-        // slower and not too much better results
-        std::set<int> already_inserted_cells[8];
-        for(int i = 0; i < 8; i++) {
-            if(!cur->children[i].is_leaf()) {
-                uninfo_refilled_nodes.push(&cur->children[i]);
-                continue;
+                // terminate subdivision on this branch
+                cur_task.node->children = nullptr; 
+                leaves.push_back(cur_task.node);
+                num_leaves++;
             }
 
-            for(int cell : cur->children[i].cells) {
-                already_inserted_cells[i].insert(cell);
-            }
+
+            // free memory
+            cur_task.points.clear();
+            cur_task.points.shrink_to_fit();
         }
 
-        for(int i = 0; i < INFORMATION_REFILLING_START_DEPTH_OFFSET && collection_start; i++) {
-            for(int cell : collection_start->cells) {
-                for(int j = 0; j < 8; j++) {
-                    if(cur->children[j].is_leaf() && !already_inserted_cells[j].count(cell)) {
-                        already_inserted_cells[j].insert(cell);
-                        cur->children[j].cells.push_back(cell);
+        refine_octree_random(univ, fallback, bounds, leaves);
+
+        // now, build the cells list
+        struct NodeComp {
+            inline bool operator()(const OctreeUncompressedNode* lhs, const OctreeUncompressedNode* rhs) const {
+                return (lhs->depth < rhs->depth);
+            }
+        };
+        std::sort(nodes_to_propagate.rbegin(), nodes_to_propagate.rend(), NodeComp());
+
+        for(auto ptr : nodes_to_propagate) {
+            auto& cur = *ptr;
+
+            if(!cur.is_leaf()) {
+                std::sort(cur.cells.begin(), cur.cells.end());
+
+                int prev_cell = cur.cells[0];
+
+                int next_idx = 1;
+                for(int i = 1; i < cur.cells.size(); i++) {
+                    if(cur.cells[i] != prev_cell) {
+                        cur.cells[next_idx] = cur.cells[i];
+                        next_idx++;
+                        prev_cell = cur.cells[i];
                     }
                 }
+                cur.cells.resize(next_idx);
+                cur.cells.shrink_to_fit();
             }
 
-            collection_start = collection_start->parent;
+            if(cur.parent) {
+                for(int cell : cur.cells) {
+                    cur.parent->cells.push_back(cell);
+                }
+            }
         }
 
-        #endif
-    }
-    information_refilling_timer.stop();
-    std::cout << "Information refilling took " << information_refilling_timer.elapsed() << " seconds." << std::endl;
+        // a possible way to remove issues with missing information in nodes
+        std::stack<OctreeUncompressedNode*> uninfo_refilled_nodes;
+        uninfo_refilled_nodes.push(&root);
+        while(!uninfo_refilled_nodes.empty()) {
+            break;
+            auto cur = uninfo_refilled_nodes.top();
+            uninfo_refilled_nodes.pop();
 
-    // now copy everything to array
-    Timer compression_timer;
-    compression_timer.start();
-    nodes.reserve(num_nodes);
-    cell_data.reserve(num_leaves);
-    compress(root);
-    compression_timer.stop();
-    std::cout << "Compression took " << compression_timer.elapsed() << std::endl;
-
-    Timer skip_traversal_building_timer;
-    skip_traversal_building_timer.start();
-
-    for(int i = 0; i < 3; i++) {
-        skip_traversal_dim[i] = (bounds.max[i] - bounds.min[i]) / SKIP_TRAVERSAL_GRID_RES;
-    }
-
-    skip_traversal_grid.resize(SKIP_TRAVERSAL_GRID_RES * SKIP_TRAVERSAL_GRID_RES * SKIP_TRAVERSAL_GRID_RES);
-    for(int i = 0; i < skip_traversal_grid.size(); i++) {
-        vec3 pos;
-
-        int ic = i;
-        for(int j = 0; j < 3; j++) {
-            int idx = ic % SKIP_TRAVERSAL_GRID_RES;
-            ic /= SKIP_TRAVERSAL_GRID_RES;
-            pos[j] = (idx + 0.5) * skip_traversal_dim[j] + bounds.min[j];
-        }
-
-        auto ptr = get_containing_node(pos, &nodes[0]);
-
-        int dist = ((char*)ptr - (char*)nodes.data()) / sizeof(*ptr);
-        skip_traversal_grid[i] = dist;
-    }
-
-    skip_traversal_building_timer.stop();
-    std::cout << "Building skip travesal took " << skip_traversal_building_timer.elapsed() << std::endl;
-
-    t.stop();
-    std::cout << "Construction took " << t.elapsed() << " seconds.\n";
-    tot_cons_time = t.elapsed();
-    //std::cout << "Expected total construction time should be " << binning_timer.elapsed() + sorting_timer.elapsed() + total_building_timer.elapsed() + 11.0 + node_propagation_timer.elapsed() + information_refilling_timer.elapsed() + compression_timer.elapsed() << " seconds." << std::endl;
-    std::cout.flush();
-
-
-
-    // now, we write the octree to disk if a path is specified
-    if(file_path.size() != 0) {
-        export_to_file(file_path);
-    }
-
-}  
-
-void OctreePartitioner::compress(const OctreeUncompressedNode& root){
-    std::queue<const OctreeUncompressedNode*> unwritten_nodes;
-    unwritten_nodes.push(&root);
-    while(!unwritten_nodes.empty()) {
-        auto cur = unwritten_nodes.front();
-        unwritten_nodes.pop();
-
-        // convert and write
-        OctreeNode compressed;
-
-        if(cur->is_leaf()) {
-            compressed.store_data(cell_data.size());
-            cell_data.push_back(std::move(cur->cells));
-            compressed.mark_leaf();
-        } else {
-            compressed.store_data(nodes.size() + 1 + unwritten_nodes.size());
+            bool should_collect_leaves = false;
             for(int i = 0; i < 8; i++) {
-                unwritten_nodes.push(&cur->children[i]);
-            }            
-        }
-
-        nodes.push_back(compressed);
-
-    }
-}
-
-OctreePartitioner::OctreePartitioner(const Universe& univ, const std::string& file_path) : fallback(univ) {
-    read_from_file(file_path);
-}
-
-OctreePartitioner::~OctreePartitioner() {
-    std::cout << "Successful octree uses:  \t" << octree_use_count << '\n';
-    std::cout << "Octree fallback uses:    \t" << fallback_use_count << '\n';
-    std::cout << "Out of bounds uses:      \t" << oob_use_count << '\n';
-    std::cout << "Again, construction time:\t" << tot_cons_time << '\n';
-}
-
-
-const OctreeNode* OctreePartitioner::get_containing_node(Position r, const OctreeNode* current) const {
-    vec3 node_dim;
-    for(int i = 0; i < 3; i++) {
-        node_dim[i] = (bounds.max[i] - bounds.min[i]) * 0.5;
-    }
-
-    vec3 center = bounds.get_center();
-    while(!current->is_leaf()) {
-        // halve the node dim
-        int idx = 0;
-        for(int i = 0; i < 3; i++) {
-            node_dim[i] *= 0.5;
-            bool less = (r[i] < center[i]);
-            center[i] += node_dim[i] * (less ? -1 : 1);
-            idx = 2 * idx + int(less);
-        }
-
-        current = &nodes[current->read_data() + idx];
-        
-    }
-
-    return current;
-}
-
-const vector<int32_t>& OctreePartitioner::get_cells(Position r, Direction u) const {
-    octree_use_count++;
-
-    if(!bounds.contains(r)) {
-        // discount this get_cells call from the stats to only focus on points within the octree
-        fallback_use_count--; 
-        octree_use_count--;
-        oob_use_count++;
-        return get_cells_fallback(r, u);
-    }
-    
-    int start = 0;
-    if(skip_traversal_grid.size() != 0) {
-        int idx = 0;
-        for(int i = 2; i >= 0; i--) {
-            int offset = int((r[i] - bounds.min[i]) / skip_traversal_dim[i]);
-            idx = SKIP_TRAVERSAL_GRID_RES * idx + offset;
-        }
-
-        start = skip_traversal_grid[idx];
-    }
-
-    auto current = get_containing_node(r, &nodes[start]);
-
-    const auto& cells = cell_data[current->read_data()];
-    if(cells.empty()) {
-        return get_cells_fallback(r, u);
-    }
-
-    return cells;
-}
-
-const vector<int32_t>& OctreePartitioner::get_cells_fallback(Position r, Direction u) const {
-    fallback_use_count++;
-    return fallback.get_cells(r, u);
-}
-
-void OctreePartitioner::export_to_file(const std::string& file_path) const {
-    std::vector<uint16_t> comp_cell_data;
-    for(const auto& v : cell_data) {
-        for(const auto& cell : v) {
-            comp_cell_data.push_back(cell);
-        }
-    }
-
-    std::fstream octree_file(file_path, std::ios::out | std::ios::binary);
-
-    // write to disk
-    #define WR_BIN(x) octree_file.write((char*)&x, sizeof(x))
-
-    for(int i = 0; i < 3; i++) {
-        WR_BIN(OMCP_CURRENT_VERSION[i]);
-    }
-    OMCPStructureType structure_type = OMCPStructureType::Octree;
-    WR_BIN(structure_type);
-
-    WR_BIN(bounds);
-
-    std::cout << "Writing " << num_nodes << " nodes and " << comp_cell_data.size() << " cells to file." << std::endl;
-
-    WR_BIN(num_nodes);
-    for(const auto& raw_node : nodes) {
-        auto node = raw_node;
-        if(node.is_leaf()) {
-            node.store_data(cell_data[node.read_data()].size());
-            node.mark_leaf();
-        }
-        WR_BIN(node);
-    }    
-
-    int num_cells = comp_cell_data.size();
-    WR_BIN(num_cells);
-    for(auto cell : comp_cell_data) {
-        WR_BIN(cell);
-    }
-
-    std::cout << "File size statistics:\n" 
-              << '\t' << "Node data  size" << ":\t" << (sizeof(nodes[0]) * nodes.size()) / (1024.0 * 1024.0) << "\tMB\n" 
-              << '\t' << "Cell data  size" << ":\t" << (sizeof(comp_cell_data[0]) * comp_cell_data.size()) / (1024.0 * 1024.0) << "\tMB\n" 
-              << '\t' << "Total file size" << ":\t" << (sizeof(nodes[0]) * nodes.size() + sizeof(comp_cell_data[0]) * comp_cell_data.size()) / (1024.0 * 1024.0) << "\tMB\n"; 
-}
-
-void OctreePartitioner::read_from_file(const std::string& file_path) {
-    std::fstream octree_file(file_path, std::ios::in | std::ios::binary);
-    #define RD_BIN(x) octree_file.read((char*)&x, sizeof(x))
-
-    int file_version[3];
-    for(int i = 0; i < 3; i++) {
-        RD_BIN(file_version[i]);
-        if(file_version[i] != OMCP_CURRENT_VERSION[i]) {
-            std::cout << "OpenMC cannot read an unsupported OMCP file version! Please note that OpenMC currently cannot read older file versions!\n";
-            std::cout.flush();
-            abort();
-        }
-    }
-
-    OMCPStructureType structure_type;
-    RD_BIN(structure_type);
-
-    if(structure_type != OMCPStructureType::Octree) {
-        std::cout << "OpenMC currently only supports octrees for the OMCP file format!\n";
-        std::cout.flush();
-        abort();
-    }
-
-    std::vector<uint16_t> comp_cell_data;
-
-    RD_BIN(bounds);
-
-    RD_BIN(num_nodes);
-    std::cout << "Octree has " << num_nodes << " nodes." << std::endl;
-
-    int num_leaves = 0;
-    nodes.resize(num_nodes);
-    for(auto& node : nodes) {
-        RD_BIN(node);
-        num_leaves++;
-    }
-    cell_data.reserve(num_leaves);
-
-    int num_cells;
-    RD_BIN(num_cells);
-    std::cout << "Octree has " << num_cells << " unique cells in leaves." << std::endl;
-    comp_cell_data.resize(num_cells);
-    for(auto& cell : comp_cell_data) {
-        RD_BIN(cell);
-    }
-
-    int next_cell_index = 0;
-    for(auto& node : nodes) {
-        if(node.is_leaf()) {
-            std::vector<int> cells;
-            cells.resize(node.read_data());
-
-            for(int& cell : cells) {
-                cell = comp_cell_data.at(next_cell_index);
-                next_cell_index++;
-            }
-
-            node.store_data(cell_data.size());
-            node.mark_leaf(); // need to do this again
-
-            cell_data.push_back(std::move(cells));
-        }
-    }
-}
-
-}
-
-/*
-parallel splitting algorithm:
-
-Successful octree uses:	120035
-Octree fallback uses:	42311
-Again, construction time:	126.471
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 23.7694 seconds.
- Universe::find_cell took 5.79047 seconds in total. (Approximately 24.361% of total execution time)
-
-
-Successful octree uses:	120181
-Octree fallback uses:	72779
-Again, construction time:	59.4075
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 26.0166 seconds.
- Universe::find_cell took 7.97749 seconds in total. (Approximately 30.663% of total execution time)
-
-
-Successful octree uses:	120036
-Octree fallback uses:	55888
-Again, construction time:	95.169
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 24.5633 seconds.
- Universe::find_cell took 6.71337 seconds in total. (Approximately 27.3309% of total execution time)
-
-
-Successful octree uses:	120072
-Octree fallback uses:	42043
-Again, construction time:	126.448
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 24.1136 seconds.
- Universe::find_cell took 5.78582 seconds in total. (Approximately 23.994% of total execution time)
-
-
-alpha 0.1, 2-2-4-8-16
-Successful octree uses:	120078
-Octree fallback uses:	42061
-Again, construction time:	123.161
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 23.8965 seconds.
- Universe::find_cell took 5.80047 seconds in total. (Approximately 24.2733% of total execution time)
-
-
-alpha 0.4
-Successful octree uses:	119959
-Octree fallback uses:	45403
-Again, construction time:	116.55
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 24.2023 seconds.
- Universe::find_cell took 6.06049 seconds in total. (Approximately 25.041% of total execution time)
-
-
-without in-place op
-Successful octree uses:	120009
-Octree fallback uses:	44386
-Again, construction time:	120.182
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 24.3735 seconds.
- Universe::find_cell took 6.08325 seconds in total. (Approximately 24.9585% of total execution time)
-
-
-after the fix (4-12-32)
-Successful octree uses:	120090
-Octree fallback uses:	44398
-Again, construction time:	119.082
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 24.2427 seconds.
- Universe::find_cell took 5.94023 seconds in total. (Approximately 24.5032% of total execution time)
-
-
-with binning
-Successful octree uses:	120001
-Octree fallback uses:	72621
-Again, construction time:	77.2118
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 25.9852 seconds.
- Universe::find_cell took 7.89101 seconds in total. (Approximately 30.3673% of total execution time)
-
-
-with no bining, constant dens 128
-Successful octree uses:	120034
-Octree fallback uses:	46379
-Again, construction time:	233.664
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 24.2725 seconds.
- Universe::find_cell took 6.091 seconds in total. (Approximately 25.0942% of total execution time)
-
-with no bining, constant dens 32
-Successful octree uses:	120184
-Octree fallback uses:	72527
-Again, construction time:	59.2701
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 26.2013 seconds.
- Universe::find_cell took 7.97431 seconds in total. (Approximately 30.4348% of total execution time)
-
-after fixing binning
-Successful octree uses:	120195
-Octree fallback uses:	72620
-Again, construction time:	76.3437
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 26.1909 seconds.
- Universe::find_cell took 7.96369 seconds in total. (Approximately 30.4063% of total execution time)
-
-affter seperation (binning)
-Successful octree uses:	120149
-Octree fallback uses:	72647
-Again, construction time:	77.1117
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 26.1907 seconds.
- Universe::find_cell took 7.91638 seconds in total. (Approximately 30.2259% of total execution time)
-
-
-after bin search
-Successful octree uses:	120131
-Octree fallback uses:	72324
-Again, construction time:	58.6078
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 25.8686 seconds.
- Universe::find_cell took 7.99754 seconds in total. (Approximately 30.9161% of total execution time)
-
-4-12-16
-Successful octree uses:	120174
-Octree fallback uses:	72944
-Again, construction time:	59.5206
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 27.0091 seconds.
- Universe::find_cell took 8.09066 seconds in total. (Approximately 29.9554% of total execution time)
-
-2-2-4-8-16
-Successful octree uses:	120140
-Octree fallback uses:	72310
-Again, construction time:	59.8562
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 26.2316 seconds.
- Universe::find_cell took 7.92618 seconds in total. (Approximately 30.2161% of total execution time)
-
-4-4-4-4-8-8
-Successful octree uses:	120151
-Octree fallback uses:	72443
-Again, construction time:	59.5919
- ============================>    PERFORMANCE STATISTICS    <============================
- Note: performance metrics only count the time spent in the transport loop and currently only support history-based transport.
- The time below is given as total wall clock seconds across all threads (e.g. 1 second of 8 threads executing is counted as 8 seconds)
- Total OpenMC execution time was 26.1838 seconds.
- Universe::find_cell took 7.90228 seconds in total. (Approximately 30.18% of total execution time)
-            bool unique = true;
-            for(int x : cells) {
-                if(x == p.cell) {
-                    unique = false;
+                if(cur->children[i].is_leaf()) {
+                    should_collect_leaves = true;
                     break;
                 }
             }
-
-            if(unique) {
-                cells.push_back(p.cell);
+            
+            auto collection_start = cur;
+            for(int i = 0; i < INFORMATION_REFILLING_START_DEPTH_OFFSET; i++) {
+                if(collection_start->parent) {
+                    collection_start = collection_start->parent;
+                }
             }
+            
+            const auto& unique_cells = collection_start->cells;
 
-Algorithm:
+            for(int i = 0; i < 8; i++) {
+                if(cur->children[i].is_leaf()) {
+                    int orig_size = cur->children[i].cells.size();
 
-construct_octree:
-    build root node
-    if number of cells in root node is less than target:
-        return
+                    auto search_start = cur->children[i].cells.begin();
+                    auto search_end   = cur->children[i].cells.begin() + orig_size;
 
-    push root node on stack
+                    cur->children[i].num_original_cells = orig_size;
 
-    while stack not empty:
-        pop node from stack and store it in variable "current"
-
-        subdivide current
-        for each child
-            count cells in child
-            if number of cells in child exceeds target
-                push child to stack
-
-search_octree:
-    current = root
-    while true
-        if current is leaf
-            break
-        else
-            for each child
-                if child contains point
-                current = child
-                break
-
-
-
-proper mt construction algo? multiple threads exec this loop:
-    finished = false
-    push root to unsplit node stack
-    while true
-        update info to console if needed
-
-        if finished
-            break
-        else if unsplit nodes remain
-            split nodes
-            push split results to unsearched node stack
-            wake threads
-        else if  unsearched nodes remain
-            search node
-            if node should be split
-                push to unsplit node stack
-            else if all other threads are sleeping
-                finished = true
-            wake threads
-        else
-            sleep
-    end while 
-
-
-
-
-
-
-
-
-
-        //std::cout << "PROCESSING..." << std::endl;
-        for(int i = 0; i < children.size(); i++) {
-            const auto& child = children[i];
-            if(child.first->cells.size() > target_cells_per_node) {
-                //omp_set_lock(&allocate_lock);
-                child.first->children = node_alloc.allocate();
-                //omp_unset_lock(&allocate_lock);
-
-                OctreeConstructionTask cur_task;
-
-                cur_task.node = child.first;
-                cur_task.points = std::move(cur_task.points);
-                cur_task.box = cur_task.node->box;
-
-                SubDivRes child_tasks;
-                auto boxes = cur_task.node->subdivide(cur_task.box);
-                for(int j = 0; j < 8; j++) {
-                    child_tasks[j].node = &cur_task.node->children[j];
-                    child_tasks[j].node->box = boxes[j];
-                    child_tasks[j].box = boxes[j];
-                    child_tasks[j].points.reserve(cur_task.points.size());
-
-                    child_tasks[j].node->depth = cur_task.node->depth + 1;
-                }
-
-                // sort points
-                for(const auto& point : cur_task.points) {
-                    child_tasks[cur_task.node->get_containing_child_index(point.pos)].points.push_back(point);
-                }
-
-                // now place them into an arr
-                for(int j = 0; j < 8; j++) {
-                    // we don't have to sort since all points were already sorted before processing loop
-                    const int NUM_UNIQUE_CELLS_MULT = 4; // prevent refinement stage from having to reallocate
-                    child_tasks[j].node->cells.reserve(NUM_UNIQUE_CELLS_MULT * cur_task.points.size());
-
-                    int prev_cell = -1;
-                    for(const auto& p : child_tasks[j].points) {
-                        if(p.cell != prev_cell) {
-                            child_tasks[j].node->cells.push_back(p.cell);
-                            prev_cell = p.cell;
+                    for(int cell : unique_cells) {
+                        if(!std::binary_search(search_start, search_end, cell)) {
+                            cur->children[i].cells.push_back(cell);
                         }
-                    }        
+                    }
+                } else {
+                    uninfo_refilled_nodes.push(&cur->children[i]);
                 }
-                // place into arr
-               // omp_set_lock(&replace_lock);
-                replacements.emplace_back(i, std::move(child_tasks));
-                //omp_unset_lock(&replace_lock);
             }
         }
 
-        //std::cout << "REPLACING..." << std::endl;
-        for(const auto& r : replacements) {
-            children[r.first].first = r.second[0].node;
-            children[r.first].second = std::move(r.second[0].points);
+        // now copy everything to array
+        nodes.reserve(num_nodes);
+        cell_data.reserve(num_leaves);
+        std::queue<const OctreeUncompressedNode*> unwritten_nodes;
+        unwritten_nodes.push(&root);
+        while(!unwritten_nodes.empty()) {
+            auto cur = unwritten_nodes.front();
+            unwritten_nodes.pop();
 
-            for(int i = 1; i < 8; i++) {
-                children.emplace_back(r.second[i].node, std::move(r.second[i].points));
+            // convert and write
+            OctreeNode compressed;
+
+            if(cur->is_leaf()) {
+                compressed.store_data(cell_data.size());
+                cell_data.push_back(std::move(cur->cells));
+                compressed.mark_as_leaf();
+            } else {
+                compressed.store_data(nodes.size() + 1 + unwritten_nodes.size());
+                for(int i = 0; i < 8; i++) {
+                    unwritten_nodes.push(&cur->children[i]);
+                }            
+            }
+
+            nodes.push_back(compressed);
+        }
+
+        refill_information();
+
+        write_message("Octree construction completed in " + std::to_string(construction_timer.elapsed()) + " seconds.", 5);
+    }  
+
+    OctreePartitioner::OctreePartitioner(const Universe& univ, const std::string& path) : fallback(univ) {
+        std::fstream octree_file(path, std::ios::in | std::ios::binary);
+        #define READ_BINARY(x) octree_file.read((char*)&x, sizeof(x))
+
+        for(int i = 0; i < 3; i++) {
+            int version;
+            READ_BINARY(version);
+            if(version != OMCP_CURRENT_VERSION[i]) {
+                fatal_error("OpenMC cannot read an unsupported OMCP file version! Please note that OpenMC currently cannot read older file versions!");
             }
         }
 
-        node_cdf.resize(children.size());
+        OMCPStructureType structure_type;
+        READ_BINARY(structure_type);
 
-
-
-
-
-
-
-    older psuedocode:
-
-        if more nodes to search points for
-            search points
-            if should split
-                push to split stack
-            else if all other threads are asleep
-                mark should exit flag
-                wake threads
-        else if more unsplit nodes
-            split nodes and push to searching stack
-            wake threads
-        else
-            sleep
-
-void find_cells_in_box(const Universe& univ, const OctreeUncompressedNode& parent, OctreeUncompressedNode& child, const AABB& box) {
-    const double NUM_CELL_SEARCH_POINT_DENSITY = 4.0;
-    const int num_search_points = (int)(box.volume() * NUM_CELL_SEARCH_POINT_DENSITY);
-
-    // get some numbers that are unique across all invocations
-    uint64_t octree_prng_seed[3] = {(uint64_t)&child, (uint64_t)child.id, (uint64_t)((child.id * 3244230925 + 432534) % 436786)};
-
-    #ifdef OCTREE_DEBUG_INFO
-    const int POINTS_LOGGING_THRESHOLD = 16384; 
-    const int POINTS_LOGGING_INCREMENT = 5000;
-    if(num_search_points > POINTS_LOGGING_THRESHOLD) {
-        LOCK_OCM();
-        std::cout << "Note: searching node " << child.id << " for " << num_search_points << " points.\n";
-        UNLOCK_OCM();
-    }
-
-    Timer t;
-    t.start();
-    #endif
-
-    std::vector<bool> skip_cell(parent.cells.size());
-    child.cells.reserve(parent.cells.size());
-    for(int i = 0; i < num_search_points; i++) {
-        #ifdef OCTREE_DEBUG_INFO
-        if(num_search_points > POINTS_LOGGING_THRESHOLD && i % POINTS_LOGGING_INCREMENT == 0 && i != 0) {
-            LOCK_OCM();
-            std::cout << "Node " << child.id << ":\tcurrently searched " << i << " out of " << num_search_points << " points (" << (100.0 * i) / num_search_points 
-                      << "%) and found " << child.cells.size() << " unique cells. ";
-            std::cout << "ETA is " << ((float)num_search_points / i * t.elapsed()) - t.elapsed() << " more seconds for this node.\n";
-            std::cout.flush(); // you may want to comment this out
-            UNLOCK_OCM();
-        }
-        #endif
-
-        vec3 rand_pos;
-
-        // gen random point
-        rand_pos.x = uniform_distribution(box.min.x, box.max.x, &octree_prng_seed[0]);
-        rand_pos.y = uniform_distribution(box.min.y, box.max.y, &octree_prng_seed[1]);
-        rand_pos.z = uniform_distribution(box.min.z, box.max.z, &octree_prng_seed[2]);
-
-        univ.find_cell_in_list(parent.cells, child.cells, skip_cell, rand_pos);
-    }
-} 
-
-
-std::vector<CellPoint> get_cell_points_uniform_dist(const Universe& univ, const UniversePartitioner& fallback, const AABB& bounds) {
-    const float NUM_CELL_SEARCH_POINT_DENSITY = 32.0f;
-    const int num_search_points = (int)(bounds.volume() * NUM_CELL_SEARCH_POINT_DENSITY);
-    std::vector<CellPoint> points_in_bounds(num_search_points);
-
-    int total_points_searched = 0;
-        
-    omp_lock_t progess_display_lock;
-    omp_init_lock(&progess_display_lock);
-
-    #pragma omp parallel for
-    for(int i = 0; i < num_search_points; i++) {
-        #pragma omp atomic
-        total_points_searched++;
-
-        if(total_points_searched % 500000 == 0 && total_points_searched != 0) {
-            omp_set_lock(&progess_display_lock);
-            std::cout << "Currently searched " << total_points_searched << " out of " << num_search_points << " points (" << (100.0 * total_points_searched) / num_search_points << "%).\tETA is " << (float(num_search_points) / total_points_searched * t.elapsed()) - t.elapsed() << "\tmore seconds for this node.\n";
-            std::cout.flush();
-            omp_unset_lock(&progess_display_lock);
+        if(structure_type != OMCPStructureType::Octree) {
+            fatal_error("OpenMC currently only supports octrees for the OMCP file format!");
         }
 
-        //auto& point = points_in_bounds[i];
+        std::vector<uint16_t> comp_cell_data;
 
-        CellPointUncompressed point;
+        READ_BINARY(bounds);
 
-        uint64_t seed[3] = {
-            (uint64_t)i, (uint64_t)(num_search_points + i), (uint64_t)(2 * num_search_points + i)
+        READ_BINARY(num_nodes);
+
+        int num_leaves = 0;
+        nodes.resize(num_nodes);
+        for(auto& node : nodes) {
+            READ_BINARY(node);
+            num_leaves++;
+        }
+        cell_data.reserve(num_leaves);
+
+        int num_cells;
+        READ_BINARY(num_cells);
+        comp_cell_data.resize(num_cells);
+        for(auto& cell : comp_cell_data) {
+            READ_BINARY(cell);
+        }
+
+        int next_cell_index = 0;
+        for(auto& node : nodes) {
+            if(node.is_leaf()) {
+                std::vector<int> cells;
+                cells.resize(node.read_data());
+
+                for(int& cell : cells) {
+                    cell = comp_cell_data.at(next_cell_index);
+                    next_cell_index++;
+                }
+
+                node.store_data(cell_data.size());
+                cell_data.push_back(std::move(cells));
+            }
+        }
+    }
+
+    void OctreePartitioner::refill_information() {
+        struct OctreeNodeExtraInfo {
+            OctreeNodeExtraInfo() : parent(nullptr), eiparent(nullptr), current(nullptr), eichildren(nullptr), children(nullptr), depth(0) {}
+
+            OctreeNode* parent;
+            OctreeNode* current;
+            OctreeNode* children;
+
+            OctreeNodeExtraInfo* eiparent;
+            OctreeNodeExtraInfo* eichildren;
+
+            std::vector<int> cells;
+
+            uint32_t depth = 0;
         };
 
-        for(int j = 0; j < 3; j++) {
-            point.pos[j] = uniform_distribution(bounds.min[j], bounds.max[j], &seed[j]);
-        }
+        std::vector<OctreeNodeExtraInfo> nodes_to_propagate(nodes.size());
+        for(int i = 0; i < nodes.size(); i++) {
+            OctreeNodeExtraInfo einode;
 
-        Direction dummy_dir{1.0, 0.0, 0.0};
-
-        const auto& possible_cells = fallback.get_cells(point.pos, dummy_dir);
-        point.cell = univ.find_cell_for_point(possible_cells, point.pos);
-
-        points_in_bounds[i].compress_from(point, bounds);
-    }
-
-    return points_in_bounds;
-}
-
-// This one is extremely difficult to tune, use the randomized optimizer instead
-void refine_octree_iterative(const Universe& univ, const UniversePartitioner& fallback, OctreeUncompressedNodeAllocator& node_alloc, OctreeUncompressedNode& root) {
-    Timer refinement_timer;
-    refinement_timer.start();
-
-    std::cout << "Refining octree via iterative refinement... note that detailed performance statistics will not be available.\n";
-    std::cout.flush();
-
-    // input paramters
-    const int NUM_ITERATIONS = 64;
-    const int NUM_BASE_SEARCH_POINTS = 1;
-    const float K_SCORE_HALF_LIFE = 9.2;
-    const float K_INIT_SCORE_MULT = 0.5;
-    const float K_CELL_DISCOVERY_REWARD = 3.0;
-    // derived parameters
-    const float K_SCORE_ALPHA = exp(-log(2.0) / K_SCORE_HALF_LIFE);
-
-    std::cout << "Using score alpha " << K_SCORE_ALPHA << '\n';
-    std::cout.flush();
-
-    std::vector<uint64_t[3]> rng_pos(NUM_THREADS);
-    for(int i = 0; i < NUM_THREADS; i++) {
-        for(int j = 0; j < 3; j++) {
-            rng_pos[i][j] = NUM_THREADS * (j + 1) + i;
-        }
-    }
-
-    auto children = store_children(root);
-
-    std::vector<float> node_score(children.size());
-    for(int i = 0; i < children.size(); i++) {
-        node_score[i] = children[i]->cells.size() * K_INIT_SCORE_MULT;
-    }
-
-    omp_lock_t loop_lock;
-    omp_init_lock(&loop_lock);
-    for(int iteration = 0; iteration < NUM_ITERATIONS; iteration++) {
-        int num_total_missing_cells = 0;
-        int num_total_searched_cells = 0;
-        int next_unproc_node = 0;
-        #pragma omp parallel for
-        for(int tid = 0; tid < NUM_THREADS; tid++) {
-            std::vector<int> untested_cells;
-            untested_cells.reserve(256);
-
-            while(true) {
-                omp_set_lock(&loop_lock);
-                int idx = next_unproc_node++;
-                omp_unset_lock(&loop_lock);
-
-                if(idx >= children.size()) {
-                    break;
-                } else {
-                    auto current = children[idx];
-
-                    if(iteration != 0) {
-                    }
-
-                    int num_missing_cells = 0;
-
-                    int num_search = std::max(NUM_BASE_SEARCH_POINTS, int(node_score[idx] + 0.5));                    
-                    num_total_searched_cells += num_search;
+            einode.current = &nodes[i];
             
-                    for(int i = 0; i < num_search; i++) {
-                        CellPointUncompressed point;
-                        for(int i = 0; i < 3; i++) {
-                            point.pos[i] = uniform_distribution(current->box.min[i], current->box.max[i], &rng_pos[tid][i]);
-                        }
-
-                        point.cell = univ.find_cell_for_point(current->cells, point.pos);
-                        if(point.cell == -1) {
-
-                            Direction dummy{0, 0, 1};
-                            const auto& possible_cells = fallback.get_cells(point.pos, dummy);
-
-                            untested_cells.clear();
-                            for(int cell : possible_cells) {
-                                if(!current->contains(cell)) {
-                                    untested_cells.push_back(cell);
-                                }
-                            }
-
-                            point.cell = univ.find_cell_for_point(untested_cells, point.pos);
-                            current->cells.push_back(point.cell);
-
-                            num_missing_cells++;
-                            num_total_missing_cells++;
-                        }
-                    }
-
-                    node_score[idx] = K_SCORE_ALPHA * K_CELL_DISCOVERY_REWARD * num_missing_cells + (1.0 - K_SCORE_ALPHA) * node_score[idx];
-                }
+            if(!nodes[i].is_leaf()) {
+                int idx = nodes[i].read_data();
+                einode.children = &nodes[idx];
+                einode.eichildren = &nodes_to_propagate[idx];
             }
         }
 
-
-        std::cout << "Refinement on iteration " << (iteration + 1) << ":\t" << num_total_searched_cells << "\tcells searched,\t" << num_total_missing_cells << "\tmising cells inserted into octree. (" << (100.0 * num_total_missing_cells) / num_total_searched_cells << "%\tdiscovery rate)\n";
-        std::cout.flush();
     }
 
-    float total = refinement_timer.elapsed();
-    std::cout << "Refinement took " << total << " seconds.\n";
-    std::cout.flush();
+    int octree_use_count = 0;
+    int fallback_use_count = 0;
+    int oob_use_count = 0;
+    OctreePartitioner::~OctreePartitioner() {
+        write_message("In bounds calls alls to OctreePartitioner::get_cell    :\t" + std::to_string(octree_use_count), 5);
+        write_message("In bounds calls to OctreePartitioner::get_cell_fallback:\t" + std::to_string(fallback_use_count), 5);
+        write_message("Out of bounds calls to OctreePartitioner::get_cell     :\t" + std::to_string(oob_use_count), 5);
+    }
+
+    const vector<int32_t>& OctreePartitioner::get_cells(Position r, Direction u) const {
+        if(!bounds.contains(r)) {
+            // discount this get_cells call from the stats to only focus on points within the octree
+            oob_use_count++;
+            return get_cells_fallback(r, u);
+        }
+        octree_use_count++;
+
+        vec3 node_dim;
+        for(int i = 0; i < 3; i++) {
+            node_dim[i] = (bounds.max[i] - bounds.min[i]) * 0.5;
+        }
+
+        vec3 center = bounds.get_center();
+        auto current = nodes[0];
+        while(!current.is_leaf()) {
+            // halve the node dim
+            int idx = 0;
+            for(int i = 0; i < 3; i++) {
+                node_dim[i] *= 0.5;
+                bool less = (r[i] < center[i]);
+                center[i] += node_dim[i] * (less ? -1 : 1);
+                idx = 2 * idx + int(less);
+            }
+
+            current = nodes[current.read_data() + idx];
+        }
+
+        const auto& cells = cell_data[current.read_data()];
+        if(cells.empty()) {
+            return get_cells_fallback(r, u);
+        }
+
+        return cells;
+    }
+
+    const vector<int32_t>& OctreePartitioner::get_cells_fallback(Position r, Direction u) const {
+        fallback_use_count++;
+        return fallback.get_cells(r, u);
+    }
+
+    void OctreePartitioner::export_to_file(const std::string& path) const {
+        std::vector<uint16_t> comp_cell_data;
+        for(const auto& v : cell_data) {
+            for(const auto& cell : v) {
+                comp_cell_data.push_back(cell);
+            }
+        }
+
+        std::fstream octree_file(path, std::ios::out | std::ios::binary);
+
+        #define WRITE_BINARY(x) octree_file.write((char*)&x, sizeof(x))
+
+        for(int i = 0; i < 3; i++) {
+            WRITE_BINARY(OMCP_CURRENT_VERSION[i]);
+        }
+        OMCPStructureType structure_type = OMCPStructureType::Octree;
+        WRITE_BINARY(structure_type);
+
+        WRITE_BINARY(bounds);
+
+        WRITE_BINARY(num_nodes);
+        for(const auto& raw_node : nodes) {
+            auto node = raw_node;
+            if(node.is_leaf()) {
+                node.store_data(cell_data[node.read_data()].size());
+            }
+            WRITE_BINARY(node);
+        }    
+
+        int num_cells = comp_cell_data.size();
+        WRITE_BINARY(num_cells);
+        for(auto cell : comp_cell_data) {
+            WRITE_BINARY(cell);
+        }
+
+        write_message("Exported octree to " + path, 5);
+    }
+
 }
-*/
